+++ instapy/comment_util.py
@@ -7,6 +7,7 @@ import emoji
 # import InstaPy modules
 from .time_util import sleep
 from .util import update_activity
+from .util import getMediaData
 from .util import add_user_to_blacklist
 from .util import click_element
 from .util import get_action_delay
@@ -180,20 +181,16 @@ def verify_mandatory_words(
 ):
     if len(mand_words) > 0 or isinstance(comments[0], dict):
         try:
-            post_desc = browser.execute_script(
-                "return window.__additionalData[Object.keys(window.__additionalData)[0]].data."
-                "graphql.shortcode_media."
-                "edge_media_to_caption.edges[0]['node']['text']"
+            post_desc = getMediaData(
+                "edge_media_to_caption.edges.0.node.text", browser
             ).lower()
 
         except Exception:
             post_desc = None
 
         try:
-            first_comment = browser.execute_script(
-                "return window.__additionalData[Object.keys(window.__additionalData)[0]].data."
-                "graphql.shortcode_media."
-                "edge_media_to_parent_comment.edges[0]['node']['text']"
+            first_comment = getMediaData(
+                "edge_media_to_parent_comment.edges.0.node.text", browser
             ).lower()
 
         except Exception:
@@ -236,7 +233,11 @@ def verify_mandatory_words(
 def get_comments_on_post(
     browser, owner, poster, amount, post_link, ignore_users, randomize, logger
 ):
+<<<<<<< HEAD
     """ Fetch comments data on posts """
+=======
+    """Fetch comments data on posts"""
+>>>>>>> upstream/master
     web_address_navigator(browser, post_link)
 
     comments = []
@@ -276,10 +277,14 @@ def get_comments_on_post(
             commenter = None
             comment = None
 
+<<<<<<< HEAD
             data = browser.execute_script(
                 "return window.__additionalData[Object.keys(window.__additionalData)].data."
                 "graphql.shortcode_media.edge_media_to_parent_comment"
             )
+=======
+            data = getMediaData("edge_media_to_parent_comment", browser)
+>>>>>>> upstream/master
             for value in data["edges"]:
                 commenter = value["node"]["owner"]["username"]
                 comment = value["node"]["text"]
@@ -333,29 +338,9 @@ def get_comments_on_post(
 
 
 def is_commenting_enabled(browser, logger):
-    """ Find out if commenting on the post is enabled """
+    """Find out if commenting on the post is enabled"""
 
-    try:
-        comments_disabled = browser.execute_script(
-            "return window.__additionalData[Object.keys(window.__additionalData)[0]].data"
-            ".graphql.shortcode_media.comments_disabled"
-        )
-
-    except WebDriverException:
-        try:
-            browser.execute_script("location.reload()")
-            update_activity(browser, state=None)
-
-            comments_disabled = browser.execute_script(
-                "return window.__additionalData[Object.keys(window.__additionalData)[0]].data"
-                ".graphql.shortcode_media.comments_disabled"
-            )
-
-        except Exception as e:
-            msg = "Failed to check comments' status for verification!\n\t{}".format(
-                str(e).encode("utf-8")
-            )
-            return False, msg
+    comments_disabled = getMediaData("comments_disabled", browser)
 
     if comments_disabled is True:
         msg = "Comments are disabled for this post."
@@ -365,13 +350,9 @@ def is_commenting_enabled(browser, logger):
 
 
 def get_comments_count(browser, logger):
-    """ Get the number of total comments in the post """
-    try:
-        comments_count = browser.execute_script(
-            "return window.__additionalData[Object.keys(window.__additionalData)[0]].data"
-            ".graphql.shortcode_media.edge_media_preview_comment.count"
-        )
+    """Get the number of total comments in the post"""
 
+<<<<<<< HEAD
     except Exception as e:
         msg = "Failed to get comments' count!\n\t{}".format(str(e).encode("utf-8"))
         return None, msg
@@ -384,16 +365,31 @@ def verify_commented_image(browser, link, owner, logger):
 
     web_address_navigator(browser, link)
 
+=======
+    comments_count = getMediaData("edge_media_preview_comment.count", browser)
+    return comments_count, "Success"
+
+
+def verify_commented_image(browser, link, owner, logger):
+    """Fetch comments data on posts to determine if already commented"""
+
+    web_address_navigator(browser, link)
+
+>>>>>>> upstream/master
     # wait for page fully load [IMPORTANT!]
     explicit_wait(browser, "PFL", [], logger, 10)
 
     try:
         commenter = None
         comment = None
+<<<<<<< HEAD
         data = browser.execute_script(
             "return window.__additionalData[Object.keys(window.__additionalData)].data."
             "graphql.shortcode_media.edge_media_to_parent_comment"
         )
+=======
+        data = getMediaData("edge_media_to_parent_comment", browser)
+>>>>>>> upstream/master
         for value in data["edges"]:
             commenter = value["node"]["owner"]["username"]
             comment = value["node"]["text"]