+++ instapy/commenters_util.py
@@ -316,7 +316,11 @@ def users_liked(browser, photo_url, amount=100, logger=None):
 
 
 def likers_from_photo(browser, amount=20, logger=None):
+<<<<<<< HEAD
     """ Get the list of users from the 'Likes' dialog of a photo """
+=======
+    """Get the list of users from the 'Likes' dialog of a photo"""
+>>>>>>> upstream/master
 
     try:
         if check_exists_by_xpath(